load custom fonts into storybook
